Change return type of UserDAO::getCount/0 to Long

+++ rumbl-domain/src/main/kotlin/ee/maytr/rumbl/domain/context/accounts/dao/UserDAO.kt
@@ -5,6 +5,6 @@ import ee.maytr.rumbl.domain.context.accounts.model.User
 interface UserDAO {
     fun get(id: Long): User?
     fun list(): List<User>
-    fun getCount(): Int
+    fun getCount(): Long
     fun create(email: String): User?
 }
