+++ rumbl-web/src/main/kotlin/ee/maytr/rumbl/web/controller/UserController.kt
@@ -24,7 +24,7 @@ class UserController(
     }
 
     @GetMapping("/users/count")
-    fun getCount(): Int {
+    fun getCount(): Long {
         return userDAO.getCount()
     }
